+++ smartcontract/programs/doublezero-serviceability/src/processors/contributor/resume.rs
@@ -65,6 +65,12 @@ pub fn process_resume_contributor(
     }
 
     let mut contributor: Contributor = Contributor::try_from(contributor_account)?;
+
+    // Only resume contributors that are currently Suspended
+    if contributor.status != ContributorStatus::Suspended {
+        return Err(DoubleZeroError::InvalidStatus.into());
+    }
+
     contributor.status = ContributorStatus::Activated;
 
     account_write(